+++ CHANGELOG.md
@@ -324,3 +324,4 @@
 - [2026-04-11 21:25] feat(storage): add bi-directional event bus handler for storage
 - [2026-04-11 21:32] fix(network): prevent double tap duplicate action in network
 - [2026-04-11 22:27] fix(analytics): prevent double tap duplicate action in analytics
+- [2026-04-11 23:03] fix(security): prevent double tap duplicate action in security